whitebox testing

+++ app/Http/Controllers/Auth/RegisterController.php
@@ -9,6 +9,8 @@ use Illuminate\Foundation\Auth\RegistersUsers;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 
+
+
 class RegisterController extends Controller
 {
     /*
@@ -52,7 +54,7 @@ class RegisterController extends Controller
         return Validator::make($data, [
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
-            'hak_akses' => ['required', 'tinyint'],
+            'hak_akses' => ['required', 'int'],
             'password' => ['required', 'string', 'min:8', 'confirmed'],
         ]);
     }
